GROUP 10 . CarDocAI/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.0.2
google-generativeai==0.3.2
google-api-python-client==2.108.0
librosa==0.10.1
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
scipy==1.11.4
soundfile==0.12.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.24.1
aiofiles==23.2.1
